Create timedeletefromADmanager
Please note that you will need to adjust the parameters of the 'Send-MailMessage' cmdlet based on your organization's email configuration. In this corrected script, the 'Credential' parameter is replaced with a new command that retrieves the user's distinguished name using the 'Get-ADUser' cmdlet and then passes it to the 'Get-Credential' cmdlet. This will prompt the user to enter their credentials when running the script.